refactor: create useScanSensor hook
